refactor to remove.ts
