+++ package.json
@@ -62,6 +62,7 @@
     ]
   },
   "dependencies": {
+    "@types/request": "^2.48.1",
     "JSONStream": "^1.2.1",
     "archiver": "^2.1.1",
     "cjson": "^0.3.1",
@@ -103,12 +104,12 @@
   },
   "devDependencies": {
     "@types/chai": "^4.1.6",
-    "@types/nock": "^9.3.0",
     "@types/chai-as-promised": "^7.1.0",
     "@types/cli-color": "^0.3.29",
     "@types/glob": "^7.1.1",
     "@types/lodash": "^4.14.118",
     "@types/mocha": "^5.2.5",
+    "@types/nock": "^9.3.0",
     "@types/node": "^10.12.0",
     "@types/sinon": "^5.0.5",
     "chai": "^4.2.0",
