+++ src/commands/open.js
@@ -38,9 +38,9 @@ var LINKS = [
   },
   { name: "Hosting", arg: "hosting", consoleUrl: "/hosting/main" },
   { name: "Hosting: Deployed Site", arg: "hosting:site" },
-  { name: "Remote Config", arg: "config", consoleUrl: "/config" },
+  { name: "RemoveRemote Config", arg: "config", consoleUrl: "/config" },
   {
-    name: "Remote Config: Conditions",
+    name: "RemoveRemote Config: Conditions",
     arg: "config:conditions",
     consoleUrl: "/config/conditions",
   },
