+++ src/test/database/remove.spec.ts
@@ -1,91 +1,9 @@
 "use strict";
 
 import { expect } from "chai";
-import * as nock from "nock";
-import * as sinon from "sinon";
 import * as pathLib from "path";
-import { SinonSandbox } from "sinon";
 
-import DatabaseRemove = require("../../database/remove");
-import helpers = require("../helpers");
-
-describe("Remote", () => {
-  const databaseRemove = new DatabaseRemove("", {
-    concurrency: 0,
-    retires: 0,
-    instance: "fake-db",
-  });
-
-  const remote = databaseRemove.remote;
-  let sandbox: SinonSandbox;
-
-  beforeEach(() => {
-    sandbox = sinon.createSandbox();
-    helpers.mockAuth(sandbox);
-  });
-
-  afterEach(() => {
-    sandbox.restore();
-    nock.cleanAll();
-  });
-
-  it("listPath should work", (done) => {
-    nock("https://fake-db.firebaseio.com")
-      .get("/.json")
-      .query({ shallow: true, limitToFirst: "50000" })
-      .reply(200, {
-        a: true,
-        x: true,
-        f: true,
-      });
-    expect(remote.listPath("/"))
-      .to.eventually.eql(["a", "x", "f"])
-      .notify(done);
-  });
-
-  it("prefetchTest should return empty", (done) => {
-    nock("https://fake-db.firebaseio.com")
-      .get("/empty/path.json")
-      .query({ timeout: "100ms" })
-      .reply(200, null);
-    expect(remote.prefetchTest("/empty/path"))
-      .to.eventually.eql("empty")
-      .notify(done);
-  });
-
-  it("prefetchTest should return large", (done) => {
-    nock("https://fake-db.firebaseio.com")
-      .get("/large/path.json")
-      .query({ timeout: "100ms" })
-      .reply(400, {
-        error:
-          "Data requested exceeds the maximum size that can be accessed with a single request.",
-      });
-    expect(remote.prefetchTest("/large/path"))
-      .to.eventually.eql("large")
-      .notify(done);
-  });
-
-  it("prefetchTest should return small", (done) => {
-    nock("https://fake-db.firebaseio.com")
-      .get("/small/path.json")
-      .query({ timeout: "100ms" })
-      .reply(200, {
-        x: "some data",
-      });
-    expect(remote.prefetchTest("/small/path"))
-      .to.eventually.eql("small")
-      .notify(done);
-  });
-
-  it("deletePath should work", () => {
-    nock("https://fake-db.firebaseio.com")
-      .delete("/a/b.json")
-      .query({ print: "silent" })
-      .reply(200, {});
-    return remote.deletePath("/a/b");
-  });
-});
+import DatabaseRemove from "../../database/remove";
 
 class TestRemote {
   public data: any;
